add isPrefilled arg (#859)
Normally, the nameEmpty error is not displayed when the field is empty. However, in the specific case where the field is automatically populated when a source file is selected, the nameEmpty error should be shown if the field is cleared after auto-completion. This behavior is controlled by the isPrefilled argument. Put it to true in the component calling use-unique-name-validation to activate this logic.

(Not: currently, the validation checks whether the entire form is dirty, rather than just tracking whether the file selection (which triggers auto-completion) has changed.)